requirements.txt
faker
pytest
requests
PyYAML
pytest-html
allure-pytest
python-dotenv